+++ client/package.json
@@ -11,7 +11,9 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@gsap/react": "^2.1.1",
     "bootstrap": "^5.3.3",
+    "gsap": "^3.12.5",
     "js-cookie": "^3.0.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
