+++ client/src/components/header/Header.jsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import './Header.css';
 import Nav from '../nav/Nav';
+import { gsap } from "gsap";
 
 export default function Header(){
