+++ client/src/pages/home/Home.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import './Home.css';
-
+import { gsap } from "gsap";
 
 export default function Home(){
