+++ client/src/pages/market/Market.css
@@ -1,10 +1,10 @@
 
-.card-container{
+.cardcontainer{
   display: flex;
   flex-wrap:wrap;
 }
 
-.card {
+.productcard {
   /* display: flex; */
   justify-content: space-around;
   flex-direction: row;
@@ -13,6 +13,7 @@
   border-radius: 8px;
   margin: 10px;
   padding: 10px;
+  opacity: 0;
   box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
   width: fit-content; 
 }
